split out serializers code into its own module
